+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-### 0.1.0
+### 0.2.0
+- initial useful release
+- added support for declarative routes via GoRoute instances
+- added support for imperative routing via `GoRoute.builder`
+- added support for setting the URL path strategy
+- added support for conditional routing
+- added support for redirection
+- added support for optional query parameters as well as positional parameters in route names
 
 ## 0.1.0
-
-* squatting on the package name (I'm not too proud to admit it)
+- squatting on the package name (I'm not too proud to admit it)